// web/src/generated/api/models/serviceBookingType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * BookWise AI API
 * HTTP API for BookWise AI.
 * OpenAPI spec version: 0.1.0
 */

export type ServiceBookingType = typeof ServiceBookingType[keyof typeof ServiceBookingType];


export const ServiceBookingType = {
  APPOINTMENT: 'APPOINTMENT',
  CLASS: 'CLASS',
} as const;